feature : 테스트 코드 작성 및 로직 점검
<description>
- 테스트
>> 아티스트 테스트
>> 앨범 테스트 중

<issue-number>

+++ src/main/java/me/gracenam/musicsampleapi/domain/soundtrack/mapper/SoundtrackMapper.java
@@ -18,7 +18,7 @@ public interface SoundtrackMapper {
 
     List<SoundtrackResponse> findByAlbumId(Long id);
 
-    int save(List<Soundtrack> soundtracks);
+    Long save(List<Soundtrack> soundtracks);
 
     void update(List<SoundtrackUpdateRequest> requests);
